requirements.txt
Django==1.9.4
django-crispy-forms==1.6.0
